+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added `moon_phases` to `world` condition
   * A list of [MoonPhases]() which must match for the condition to be met
+* Changed `world_whitelist` to `conditions` in Attributes, Potion Effects and Items.
 
 ## 4.4.0
 * Updated to 1.19.4